extract getTableExpressions variants into own file
