+++ dbms/src/Interpreters/DatabaseAndTableWithAlias.cpp
@@ -1,6 +1,8 @@
 #include <Interpreters/DatabaseAndTableWithAlias.h>
 #include <Interpreters/IdentifierSemantic.h>
 #include <Interpreters/Context.h>
+#include <Interpreters/getTableExpressions.h>
+
 #include <Common/typeid_cast.h>
 
 #include <Parsers/IAST.h>
@@ -13,9 +15,6 @@ namespace DB
 {
 
 NameSet removeDuplicateColumns(NamesAndTypesList & columns);
-
-struct ASTTableExpression;
-NamesAndTypesList getNamesAndTypeListFromTableExpression(const ASTTableExpression & table_expression, const Context & context);
 
 DatabaseAndTableWithAlias::DatabaseAndTableWithAlias(const ASTIdentifier & identifier, const String & current_database)
 {
@@ -70,44 +69,9 @@ String DatabaseAndTableWithAlias::getQualifiedNamePrefix(bool with_dot) const
     return (!alias.empty() ? alias : table) + (with_dot ? "." : "");
 }
 
-std::vector<const ASTTableExpression *> getSelectTablesExpression(const ASTSelectQuery & select_query)
-{
-    if (!select_query.tables())
-        return {};
-
-    std::vector<const ASTTableExpression *> tables_expression;
-
-    for (const auto & child : select_query.tables()->children)
-    {
-        const auto * tables_element = child->as<ASTTablesInSelectQueryElement>();
-
-        if (tables_element->table_expression)
-            tables_expression.emplace_back(tables_element->table_expression->as<ASTTableExpression>());
-    }
-
-    return tables_expression;
-}
-
-static const ASTTableExpression * getTableExpression(const ASTSelectQuery & select, size_t table_number)
-{
-    if (!select.tables())
-        return {};
-
-    const auto & tables_in_select_query = select.tables()->as<ASTTablesInSelectQuery &>();
-    if (tables_in_select_query.children.size() <= table_number)
-        return {};
-
-    const auto & tables_element = tables_in_select_query.children[table_number]->as<ASTTablesInSelectQueryElement &>();
-
-    if (!tables_element.table_expression)
-        return {};
-
-    return tables_element.table_expression->as<ASTTableExpression>();
-}
-
 std::vector<DatabaseAndTableWithAlias> getDatabaseAndTables(const ASTSelectQuery & select_query, const String & current_database)
 {
-    std::vector<const ASTTableExpression *> tables_expression = getSelectTablesExpression(select_query);
+    std::vector<const ASTTableExpression *> tables_expression = getTableExpressions(select_query);
 
     std::vector<DatabaseAndTableWithAlias> database_and_table_with_aliases;
     database_and_table_with_aliases.reserve(tables_expression.size());
@@ -129,45 +93,6 @@ std::optional<DatabaseAndTableWithAlias> getDatabaseAndTable(const ASTSelectQuer
         return {};
 
     return DatabaseAndTableWithAlias(database_and_table_name);
-}
-
-ASTPtr extractTableExpression(const ASTSelectQuery & select, size_t table_number)
-{
-    if (const ASTTableExpression * table_expression = getTableExpression(select, table_number))
-    {
-        if (table_expression->database_and_table_name)
-            return table_expression->database_and_table_name;
-
-        if (table_expression->table_function)
-            return table_expression->table_function;
-
-        if (table_expression->subquery)
-            return table_expression->subquery->children[0];
-    }
-
-    return nullptr;
-}
-
-std::vector<TableWithColumnNames> getDatabaseAndTablesWithColumnNames(const ASTSelectQuery & select_query, const Context & context)
-{
-    std::vector<TableWithColumnNames> tables_with_columns;
-
-    if (select_query.tables() && !select_query.tables()->children.empty())
-    {
-        String current_database = context.getCurrentDatabase();
-
-        for (const ASTTableExpression * table_expression : getSelectTablesExpression(select_query))
-        {
-            DatabaseAndTableWithAlias table_name(*table_expression, current_database);
-
-            NamesAndTypesList names_and_types = getNamesAndTypeListFromTableExpression(*table_expression, context);
-            removeDuplicateColumns(names_and_types);
-
-            tables_with_columns.emplace_back(std::move(table_name), names_and_types.getNames());
-        }
-    }
-
-    return tables_with_columns;
 }
 
 }
