+++ dbms/src/Storages/LiveView/StorageLiveView.cpp
@@ -37,7 +37,7 @@ limitations under the License. */
 #include <Storages/StorageFactory.h>
 #include <Parsers/ASTTablesInSelectQuery.h>
 #include <Parsers/ASTSubquery.h>
-#include <Interpreters/DatabaseAndTableWithAlias.h>
+#include <Interpreters/getTableExpressions.h>
 #include <Interpreters/AddDefaultDatabaseVisitor.h>
 
 namespace DB
